Storm-dft: export 'parameters' field only if they are present
